test(core): back to directory based tests
