+++ packages/core/test/projects/basic/Empty.abi.json
