+++ packages/core/test/projects/basic/schema.graphql
@@ -1,0 +1,3 @@
+type Entity @entity {
+  id: ID!
+}
