+++ packages/core/test/projects/basic/src/index.ts
@@ -1,0 +1,5 @@
+import { ponder } from "../generated";
+
+ponder.on("Source:Event", async ({ event, context }) => {
+  return;
+});
