+++ packages/core/test/projects/ens/BaseRegistrarImplementation.abi.json
