+++ packages/core/test/projects/ens/schema.graphql
@@ -1,0 +1,12 @@
+type EnsNft @entity {
+  id: ID!
+  labelHash: String!
+  owner: Account!
+  transferredAt: Int!
+}
+
+type Account @entity {
+  id: ID!
+  tokens: [EnsNft!]! @derivedFrom(field: "owner")
+  lastActive: Int!
+}
