+++ packages/core/tsconfig.json
@@ -17,5 +17,5 @@
     }
   },
   "include": ["src", "test"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "test/projects"]
 }
